Revert test refactoring
